fix: Address all critical issues from feedback
CRITICAL FIXES (🔴):
1. Created app/auth.py with JWT validation for Keycloak
2. Fixed broken query in get_session (select from relationship)
3. Replaced .x/.y with to_shape() for GeoAlchemy2 geometry objects
4. Added is_finished()/is_completed() methods to models
5. Added separate Keycloak database in docker-compose
6. Fixed init_db.sql filename reference

IMPORTANT FIXES (🟡):
7. Separated networks: public_net + internal_net (isolated)
8. Fixed finish_session re-query bug
9. Added view_extent_km to StartSessionRequest schema
10. Changed average_score from Integer to Float in GameSession
11. Created migration 003 to fix difficulty constraint (1-7)
12. Removed default.conf, kept only locationking.ru.conf
13. Updated authentication in routers to use new auth module

OTHER IMPROVEMENTS:
- Updated all routers to use proper authentication
- Fixed geometry coordinate extraction with to_shape()
- Added proper network isolation for security
- All fixes based on detailed feedback analysis

+++ backend/app/models/game_session.py
@@ -38,7 +38,7 @@ class GameSession(Base):
     rounds_done: Mapped[int] = mapped_column(SmallInteger, default=0)
     total_score: Mapped[int] = mapped_column(Integer, default=0)
     average_score: Mapped[Optional[float]] = mapped_column(
-        Integer, default=0
+        Float, default=0.0
     )
     best_round_score: Mapped[int] = mapped_column(Integer, default=0)
     worst_round_score: Mapped[int] = mapped_column(Integer, default=0)
